Add type key
This commit will allow the use of the 'type' keywork in a routes YAML
file. This will contain information about the type of page the state
routes to, which can be used to determine how to display the page.

If a route is defined as having a type, The generated routes object will
contain a 'type' key at the same level as the 'on' key.

Add unit tests for type key

This commit adds some unit tests for the type key to ensure only valid
values are entered.

Moved 'start' and 'summary' up a level in the routes json object

Added unit test for summary key - should throw error if summary key does
not exist in the sections